requirement.txt
absl-py==2.1.0
contourpy==1.2.0
cvbase==0.5.5
cycler==0.12.1
einops==0.7.0
fonttools==4.49.0
fsspec==2024.2.0
grpcio==1.62.1
h5py==3.10.0
joblib==1.3.2
kiwisolver==1.4.5
loguru==0.7.2
markdown==3.5.2
matplotlib==3.8.3
msgpack==1.0.8
msgpack-numpy==0.4.8
multimethod==1.11.2
onnx==1.15.0
opencv-python==4.9.0.80
packaging==23.2
pandas==2.2.1
protobuf==5.26.0
psutil==5.9.8
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
pyzmq==25.1.2
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
six==1.16.0
tabulate==0.9.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorboardx==2.6.2.2
tensorpack==0.11
termcolor==2.4.0
terminaltables==3.1.10
threadpoolctl==3.3.0
toml==0.10.2
torchpack==0.0.3
tqdm==4.66.2
tzdata==2024.1
werkzeug==3.0.1
yacs==0.1.8
